+++ astropy/table/table.py
@@ -18,6 +18,7 @@ from astropy.utils import isiterable, ShapedLikeNDArray
 from astropy.utils.console import color_print
 from astropy.utils.metadata import MetaData
 from astropy.utils.data_info import BaseColumnInfo, MixinInfo, ParentDtypeInfo, DataInfo
+from astropy.utils.decorators import format_doc
 
 from . import groups
 from .pprint import TableFormatter
@@ -37,6 +38,82 @@ __doctest_skip__ = ['Table.read', 'Table.write',
 
 __doctest_requires__ = {'*pandas': ['pandas']}
 
+_pprint_docs = """
+    {__doc__}
+
+    Parameters
+    ----------
+    max_lines : int or `None`
+        Maximum number of lines in table output.
+
+    max_width : int or `None`
+        Maximum character width of output.
+
+    show_name : bool
+        Include a header row for column names. Default is True.
+
+    show_unit : bool
+        Include a header row for unit.  Default is to show a row
+        for units only if one or more columns has a defined value
+        for the unit.
+
+    show_dtype : bool
+        Include a header row for column dtypes. Default is True.
+
+    align : str or list or tuple or `None`
+        Left/right alignment of columns. Default is right (None) for all
+        columns. Other allowed values are '>', '<', '^', and '0=' for
+        right, left, centered, and 0-padded, respectively. A list of
+        strings can be provided for alignment of tables with multiple
+        columns.
+    """
+
+_pformat_docs = """
+    {__doc__}
+
+    Parameters
+    ----------
+    max_lines : int or `None`
+        Maximum number of rows to output
+
+    max_width : int or `None`
+        Maximum character width of output
+
+    show_name : bool
+        Include a header row for column names. Default is True.
+
+    show_unit : bool
+        Include a header row for unit.  Default is to show a row
+        for units only if one or more columns has a defined value
+        for the unit.
+
+    show_dtype : bool
+        Include a header row for column dtypes. Default is True.
+
+    html : bool
+        Format the output as an HTML table. Default is False.
+
+    tableid : str or `None`
+        An ID tag for the table; only used if html is set.  Default is
+        "table{id}", where id is the unique integer id of the table object,
+        id(self)
+
+    align : str or list or tuple or `None`
+        Left/right alignment of columns. Default is right (None) for all
+        columns. Other allowed values are '>', '<', '^', and '0=' for
+        right, left, centered, and 0-padded, respectively. A list of
+        strings can be provided for alignment of tables with multiple
+        columns.
+
+    tableclass : str or list of str or `None`
+        CSS classes for the table; only used if html is set.  Default is
+        None.
+
+    Returns
+    -------
+    lines : list
+        Formatted table as a list of strings.
+    """
 
 class TableReplaceWarning(UserWarning):
     """
@@ -189,16 +266,17 @@ class TableColumns(OrderedDict):
 class Table:
     """A class to represent tables of heterogeneous data.
 
-    `Table` provides a class for heterogeneous tabular data, making use of a
-    `numpy` structured array internally to store the data values.  A key
-    enhancement provided by the `Table` class is the ability to easily modify
-    the structure of the table by adding or removing columns, or adding new
-    rows of data.  In addition table and column metadata are fully supported.
+    `~astropy.table.Table` provides a class for heterogeneous tabular data,
+    making use of a `numpy` structured array internally to store the data
+    values.  A key enhancement provided by the `~astropy.table.Table` class is
+    the ability to easily modify the structure of the table by adding or
+    removing columns, or adding new rows of data.  In addition table and column
+    metadata are fully supported.
 
-    `Table` differs from `~astropy.nddata.NDData` by the assumption that the
-    input data consists of columns of homogeneous data, where each column
-    has a unique identifier and may contain additional metadata such as the
-    data unit, format, and description.
+    `~astropy.table.Table` differs from `~astropy.nddata.NDData` by the
+    assumption that the input data consists of columns of homogeneous data,
+    where each column has a unique identifier and may contain additional
+    metadata such as the data unit, format, and description.
 
     See also: http://docs.astropy.org/en/stable/table/
 
@@ -946,6 +1024,7 @@ class Table:
         # unit set).
         return has_info_class(col, MixinInfo) and not has_info_class(col, QuantityInfo)
 
+    @format_doc(_pprint_docs)
     def pprint(self, max_lines=None, max_width=None, show_name=True,
                show_unit=None, show_dtype=False, align=None):
         """Print a formatted string representation of the table.
@@ -960,31 +1039,6 @@ class Table:
         The same applies for max_width except the configuration item is
         ``astropy.conf.max_width``.
 
-        Parameters
-        ----------
-        max_lines : int
-            Maximum number of lines in table output.
-
-        max_width : int or `None`
-            Maximum character width of output.
-
-        show_name : bool
-            Include a header row for column names. Default is True.
-
-        show_unit : bool
-            Include a header row for unit.  Default is to show a row
-            for units only if one or more columns has a defined value
-            for the unit.
-
-        show_dtype : bool
-            Include a header row for column dtypes. Default is True.
-
-        align : str or list or tuple or `None`
-            Left/right alignment of columns. Default is right (None) for all
-            columns. Other allowed values are '>', '<', '^', and '0=' for
-            right, left, centered, and 0-padded, respectively. A list of
-            strings can be provided for alignment of tables with multiple
-            columns.
         """
         lines, outs = self.formatter._pformat_table(self, max_lines, max_width,
                                                     show_name=show_name, show_unit=show_unit,
@@ -999,6 +1053,20 @@ class Table:
                 color_print(line, 'red')
             else:
                 print(line)
+
+    @format_doc(_pprint_docs)
+    def pprint_all(self, max_lines=-1, max_width=-1, show_name=True,
+               show_unit=None, show_dtype=False, align=None):
+        """Print a formatted string representation of the entire table.
+
+        This method is the same as `astropy.table.Table.pprint` except that
+        the default ``max_lines`` and ``max_width`` are both -1 so that by
+        default the entire table is printed instead of restricting to the size
+        of the screen terminal.
+
+        """
+        return self.pprint(max_lines, max_width, show_name,
+               show_unit, show_dtype, align)
 
     def _make_index_row_display_table(self, index_row_name):
         if index_row_name not in self.columns:
@@ -1152,6 +1220,7 @@ class Table:
         else:
             br.open(urljoin('file:', pathname2url(path)))
 
+    @format_doc(_pformat_docs, id="{id}")
     def pformat(self, max_lines=None, max_width=None, show_name=True,
                 show_unit=None, show_dtype=False, html=False, tableid=None,
                 align=None, tableclass=None):
@@ -1168,49 +1237,6 @@ class Table:
         The same applies for ``max_width`` except the configuration item  is
         ``astropy.conf.max_width``.
 
-        Parameters
-        ----------
-        max_lines : int or `None`
-            Maximum number of rows to output
-
-        max_width : int or `None`
-            Maximum character width of output
-
-        show_name : bool
-            Include a header row for column names. Default is True.
-
-        show_unit : bool
-            Include a header row for unit.  Default is to show a row
-            for units only if one or more columns has a defined value
-            for the unit.
-
-        show_dtype : bool
-            Include a header row for column dtypes. Default is True.
-
-        html : bool
-            Format the output as an HTML table. Default is False.
-
-        tableid : str or `None`
-            An ID tag for the table; only used if html is set.  Default is
-            "table{id}", where id is the unique integer id of the table object,
-            id(self)
-
-        align : str or list or tuple or `None`
-            Left/right alignment of columns. Default is right (None) for all
-            columns. Other allowed values are '>', '<', '^', and '0=' for
-            right, left, centered, and 0-padded, respectively. A list of
-            strings can be provided for alignment of tables with multiple
-            columns.
-
-        tableclass : str or list of str or `None`
-            CSS classes for the table; only used if html is set.  Default is
-            None.
-
-        Returns
-        -------
-        lines : list
-            Formatted table as a list of strings.
-
         """
 
         lines, outs = self.formatter._pformat_table(
@@ -1222,6 +1248,29 @@ class Table:
             lines.append('Length = {0} rows'.format(len(self)))
 
         return lines
+
+    @format_doc(_pformat_docs, id="{id}")
+    def pformat_all(self, max_lines=-1, max_width=-1, show_name=True,
+                show_unit=None, show_dtype=False, html=False, tableid=None,
+                align=None, tableclass=None):
+        """Return a list of lines for the formatted string representation of
+        the entire table.
+
+        If no value of ``max_lines`` is supplied then the height of the
+        screen terminal is used to set ``max_lines``.  If the terminal
+        height cannot be determined then the default is taken from the
+        configuration item ``astropy.conf.max_lines``.  If a negative
+        value of ``max_lines`` is supplied then there is no line limit
+        applied.
+
+        The same applies for ``max_width`` except the configuration item  is
+        ``astropy.conf.max_width``.
+
+        """
+
+        return self.pformat(max_lines, max_width, show_name,
+                show_unit, show_dtype, html, tableid,
+                align, tableclass)
 
     def more(self, max_lines=None, max_width=None, show_name=True,
              show_unit=None, show_dtype=False):
@@ -2447,7 +2496,7 @@ class Table:
 
         # use index sorted order if possible
         if keys is not None:
-            index = get_index(self, self[keys])
+            index = get_index(self, names=keys)
             if index is not None:
                 return index.sorted_data()
 
@@ -2458,7 +2507,7 @@ class Table:
             kwargs['kind'] = kind
 
         if keys:
-            data = self[keys].as_array()
+            data = self.as_array(names=keys)
         else:
             data = self.as_array()
 
@@ -2507,7 +2556,7 @@ class Table:
             keys = [keys]
 
         indexes = self.argsort(keys)
-        sort_index = get_index(self, self[keys])
+        sort_index = get_index(self, names=keys)
         if sort_index is not None:
             # avoid inefficient relabelling of sorted index
             prev_frozen = sort_index._frozen
@@ -2582,7 +2631,7 @@ class Table:
 
         Returns
         -------
-        out : `Table`
+        out : `~astropy.table.Table`
             Table corresponding to file contents
 
         Notes
@@ -2712,25 +2761,25 @@ class Table:
         """
         Group this table by the specified ``keys``
 
-        This effectively splits the table into groups which correspond to
-        unique values of the ``keys`` grouping object.  The output is a new
-        `TableGroups` which contains a copy of this table but sorted by row
-        according to ``keys``.
+        This effectively splits the table into groups which correspond to unique
+        values of the ``keys`` grouping object.  The output is a new
+        `~astropy.table.TableGroups` which contains a copy of this table but
+        sorted by row according to ``keys``.
 
         The ``keys`` input to `group_by` can be specified in different ways:
 
           - String or list of strings corresponding to table column name(s)
           - Numpy array (homogeneous or structured) with same length as this table
-          - `Table` with same length as this table
+          - `~astropy.table.Table` with same length as this table
 
         Parameters
         ----------
-        keys : str, list of str, numpy array, or `Table`
+        keys : str, list of str, numpy array, or `~astropy.table.Table`
             Key grouping object
 
         Returns
         -------
-        out : `Table`
+        out : `~astropy.table.Table`
             New table with groups set
         """
         return groups.table_group_by(self, keys)
@@ -2841,7 +2890,7 @@ class Table:
                     tbl[col.info.name] = new_col
 
             # Convert the table to one with no mixins, only Column objects.
-            encode_tbl = serialize._represent_mixins_as_columns(tbl)
+            encode_tbl = serialize.represent_mixins_as_columns(tbl)
             return encode_tbl
 
         tbl = _encode_mixins(self)
@@ -2876,7 +2925,7 @@ class Table:
     @classmethod
     def from_pandas(cls, dataframe, index=False):
         """
-        Create a `Table` from a :class:`pandas.DataFrame` instance
+        Create a `~astropy.table.Table` from a :class:`pandas.DataFrame` instance
 
         In addition to converting generic numeric or string columns, this supports
         conversion of pandas Date and Time delta columns to `~astropy.time.Time`
@@ -2891,8 +2940,8 @@ class Table:
 
         Returns
         -------
-        table : `Table`
-            A `Table` (or subclass) instance
+        table : `~astropy.table.Table`
+            A `~astropy.table.Table` (or subclass) instance
 
         Raises
         ------
@@ -2991,12 +3040,12 @@ class Table:
 class QTable(Table):
     """A class to represent tables of heterogeneous data.
 
-    `QTable` provides a class for heterogeneous tabular data which can be
-    easily modified, for instance adding columns or new rows.
+    `~astropy.table.QTable` provides a class for heterogeneous tabular data
+    which can be easily modified, for instance adding columns or new rows.
 
-    The `QTable` class is identical to `Table` except that columns with an
-    associated ``unit`` attribute are converted to `~astropy.units.Quantity`
-    objects.
+    The `~astropy.table.QTable` class is identical to `~astropy.table.Table`
+    except that columns with an associated ``unit`` attribute are converted to
+    `~astropy.units.Quantity` objects.
 
     See also: